feat: introduce useWidgetMode hook and update theme handling
