+++ apps/cowswap-frontend/src/theme/types.ts
@@ -1,9 +1,7 @@
 import type { CowProtocolTheme } from '@cowprotocol/ui'
 
 declare module 'styled-components' {
-  export interface CowSwapDefaultTheme extends CowProtocolTheme {
-    isInjectedWidgetMode: boolean
-  }
+  export interface CowSwapDefaultTheme extends CowProtocolTheme {}
 
   export interface DefaultTheme extends CowSwapDefaultTheme {}
 }
