+++ apps/explorer/src/theme/types.ts
@@ -11,14 +11,6 @@ declare module 'styled-components' {
   export interface ExplorerTheme extends Colors, Fonts {
     // theming
     mode: 'light' | 'dark'
-    isInjectedWidgetMode: boolean
-    isStandalone: boolean
-    isWidget: boolean
-    isIframe: boolean
-    isLimitOrder: boolean
-    isSwapOrder: boolean
-    isTwapOrder: boolean
-    isAdvancedOrder: boolean
   }
 
   export interface DefaultTheme extends CowProtocolTheme, ExplorerTheme {}
