+++ libs/common-hooks/src/index.ts
@@ -17,3 +17,4 @@ export * from './useWindowSize'
 export * from './useCopyClipboard'
 export * from './useFeatureFlags'
 export * from './useGasLimitHooks'
+export * from './useWidgetMode'
